requirements.txt
fastapi
uvicorn
pytest
flake8
httpx
python-jose
passlib
bcrypt
python-multipart